change from isdeleted => status Project insert manager & worker = list<Long>

+++ src/main/java/com/ntv/ntvcons_backend/entities/RequestModels/ShowRequestModel.java
@@ -19,5 +19,5 @@ public class ShowRequestModel {
 
     private Long createdBy, updatedBy;
     private LocalDateTime createdAt, updatedAt;
-    private Double totalPage;
+    private Integer totalPage;
 }
